Implemented resource load from resource.ini file

+++ src/kml/Starter.java
@@ -1,8 +1,10 @@
 package kml;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.InputStream;
+import java.util.Properties;
 
 /**
  * @website https://krothium.com
@@ -31,6 +33,10 @@ public class Starter {
     }
     private static boolean existsResource() {
         try {
+            File custom = new File("resource.ini");
+            if (custom.exists() && custom.isFile()){
+                return true;
+            }
             InputStream in = Starter.class.getResourceAsStream("/resource.jar");
             return in != null;
         } catch (Exception ex){
@@ -38,30 +44,43 @@ public class Starter {
         }
     }
     private static void bootFromResource(String[] passedArgs) {
-        InputStream in = Starter.class.getResourceAsStream("/resource.jar");
-        File workingDir = Utils.getWorkingDirectory();
-        if (!workingDir.exists() || !workingDir.isDirectory()){
-            workingDir.mkdirs();
-        }
-        File resource = new File(workingDir + File.separator + "resource.jar");
-        boolean copy = true;
-        if (resource.exists() && resource.isFile()){
-            copy = resource.delete();
-        }
-        if (copy) {
+        File custom = new File("resource.ini");
+        if (custom.exists() && custom.isFile()){
+            Properties p = new Properties();
             try {
-                FileOutputStream out = new FileOutputStream(resource);
-                byte[] buffer = new byte[4096];
-                int read;
-                while ((read = in.read(buffer)) != -1) {
-                    out.write(buffer, 0, read);
-                }
-                in.close();
-                out.close();
-            } catch (Exception ex) {
+                FileInputStream fin = new FileInputStream(custom);
+                p.load(fin);
+                File resource = new File(p.getProperty("path"));
+                StubLauncher.load(resource, passedArgs);
+            } catch (Exception ex){
                 ex.printStackTrace();
             }
+        } else {
+            InputStream in = Starter.class.getResourceAsStream("/resource.jar");
+            File workingDir = Utils.getWorkingDirectory();
+            if (!workingDir.exists() || !workingDir.isDirectory()){
+                workingDir.mkdirs();
+            }
+            File resource = new File(workingDir + File.separator + "resource.jar");
+            boolean copy = true;
+            if (resource.exists() && resource.isFile()){
+                copy = resource.delete();
+            }
+            if (copy) {
+                try {
+                    FileOutputStream out = new FileOutputStream(resource);
+                    byte[] buffer = new byte[4096];
+                    int read;
+                    while ((read = in.read(buffer)) != -1) {
+                        out.write(buffer, 0, read);
+                    }
+                    in.close();
+                    out.close();
+                } catch (Exception ex) {
+                    ex.printStackTrace();
+                }
+            }
+            StubLauncher.load(resource, passedArgs);
         }
-        StubLauncher.load(resource, passedArgs);
     }
 }
